python/requirements.txt
# requirements.txt
# Core dependencies for RAG Anywhere

# Vector operations and search
numpy>=2.3.0
sentence-transformers>=5.1.2
transformers>=4.48.0
accelerate>=1.2.1
gliner>=0.2.22
faiss-cpu>=1.13.0

# Configuration and data handling
pyyaml>=6.0.3
pydantic>=2.12.4
pydantic-settings>=2.12.0

# Document loading
pypdf>=6.3.0
python-docx>=1.2.0

# Development and testing
pytest>=9.0.1
pytest-cov>=7.0.0
black>=25.11.0
ruff>=0.14.6
mypy>=1.18.2
